feat: update PopupState
