+++ src/containers/Home/Announcement/components/AnnouncementTable.tsx
@@ -83,7 +83,7 @@ const AnnouncementTable: FC<Props> = ({
                     deleteAnnouncementById({ variables: { id: curId } })
                   }
                 >
-                  <DeleteOutline />
+                  <DeleteOutline className={classes.editIcon} />
                 </ConfirmPoper>
               </FormControl>
