--- src/containers/Settings/Account/Account.tsx
@@ -1,7 +1,0 @@
-import React, { FC } from 'react'
-
-const Account: FC = () => {
-  return <div>account</div>
-}
-
-export default Account
