+++ src/containers/Settings/Global/Global.tsx
@@ -1,0 +1,7 @@
+import React, { FC } from 'react'
+
+const Global: FC = () => {
+  return <div>Global</div>
+}
+
+export default Global
