+++ src/containers/Settings/Security/components/ChangePassword/ChangePassword.tsx
@@ -5,7 +5,7 @@ import { useFormik } from 'formik'
 import { useMutation } from '@apollo/react-hooks'
 import { TextField, Button } from '@material-ui/core'
 import { CHANGE_PASSWORD } from '../../typeDefs'
-import SettingItemWrapper from 'src/components/SettingItemWrapper/SettingItemWrapper'
+import SettingItemWrapper from 'src/containers/Settings/components/SettingItemWrapper/SettingItemWrapper'
 import styles from './changePassword.module.scss'
 import { logout } from 'src/shared/utils'
 import { PASSWORD_REGEXP } from 'src/shared/constants'
